Выполнение задание на организацию кода. Модули
